+++ pypit/arms.py
@@ -8,7 +8,6 @@ import os
 from pypit import msgs
 from pypit import arparse as settings
 from pypit import arload
-from pypit import armbase
 from pypit import arproc
 from pypit.core import arprocimg
 from pypit import armasters
@@ -26,6 +25,7 @@ from pypit import fluxspec
 from pypit import traceslits
 from pypit import traceimage
 from pypit import wavecalib
+from pypit import waveimage
 
 from pypit import ardebug as debugger
 
@@ -250,7 +250,8 @@ def ARMS(fitstbl, setup_dict, reuseMaster=False, reloadMaster=True, sciexp=None)
                 wv_maskslits = calib_dict[setup]['wvmask']
             elif settings.argflag["reduce"]["calibrate"]["wavelength"] == "pixel":
                 msgs.info("A wavelength calibration will not be performed")
-                pass
+                wv_calib = None
+                wv_maskslits = np.zeros_like(slf._maskslits[det-1], dtype=bool)
             else:
                 # Setup up the settings (will be Refactored with settings)
                 tmp = dict(calibrate=settings.argflag['arc']['calibrate'], masters=settings.argflag['reduce']['masters'])
@@ -303,7 +304,7 @@ def ARMS(fitstbl, setup_dict, reuseMaster=False, reloadMaster=True, sciexp=None)
                                                       wv_calib=wv_calib)
                     waveTilts.save_master()
                 else:
-                    wt_maskslits = np.zeros(len(slf._maskslits[det-1]), dtype=bool)
+                    wt_maskslits = np.zeros_like(slf._maskslits[det-1], dtype=bool)
                 # Save
                 calib_dict[setup]['tilts'] = mstilts
                 calib_dict[setup]['wtmask'] = wt_maskslits
@@ -354,9 +355,29 @@ def ARMS(fitstbl, setup_dict, reuseMaster=False, reloadMaster=True, sciexp=None)
 
             ###############
             # Generate/load a master wave frame
-            update = slf.MasterWave(det, wv_calib, mstilts)
-            if update and reuseMaster:
-                armbase.UpdateMasters(sciexp, sc, det, ftype="arc", chktype="wave")
+
+            if 'wave' in calib_dict[setup].keys():
+                mswave = calib_dict[setup]['wave']
+            else:
+                if settings.argflag["reduce"]["calibrate"]["wavelength"] == "pixel":
+                    mswave = mstilts * (mstilts.shape[0]-1.0)
+                else:
+                    # Settings
+                    wvimg_settings = dict(masters=settings.argflag['reduce']['masters'].copy())
+                    wvimg_settings['masters']['directory'] = settings.argflag['run']['directory']['master']+'_'+ settings.argflag['run']['spectrograph']
+                    # Instantiate
+                    waveImage = waveimage.WaveImage(mstilts, wv_calib, settings=wvimg_settings,
+                                                setup=setup, maskslits=slf._maskslits[det-1],
+                                                slitpix=traceSlits.slitpix)
+                    # Attempt to load master
+                    mswave = waveImage.master()
+                    if mswave is None:
+                        mswave = waveImage._build_wave()
+                    # Save to hard-drive
+                    waveImage.save_master(mswave, steps=waveImage.steps)
+                # Save internally
+                calib_dict[setup]['wave'] = mswave
+
 
             ###############
             # Load the science frame and from this generate a Poisson error frame
@@ -368,7 +389,7 @@ def ARMS(fitstbl, setup_dict, reuseMaster=False, reloadMaster=True, sciexp=None)
             # Extract
             msgs.info("Processing science frame")
             arproc.reduce_multislit(slf, mstilts, sciframe, msbpm, datasec_img, scidx, fitstbl, det,
-                                    mspixelflatnrm=mspixflatnrm, slitprof=slitprof)
+                                    mswave, mspixelflatnrm=mspixflatnrm, slitprof=slitprof)
 
 
             ######################################################
@@ -390,7 +411,7 @@ def ARMS(fitstbl, setup_dict, reuseMaster=False, reloadMaster=True, sciexp=None)
                 # Reduce
                 msgs.info("Processing standard frame")
                 arproc.reduce_multislit(stdslf, mstilts, stdframe, msbpm, datasec_img, std_idx, fitstbl, det,
-                                        standard=True, mspixelflatnrm=mspixflatnrm, slitprof=slitprof)
+                                        mswave, mspixelflatnrm=mspixflatnrm, standard=True, slitprof=slitprof)
                 # Finish
                 stdslf.extracted[det-1] = True