+++ CHANGES.rst
@@ -153,13 +153,16 @@
 - New pypit_chk_tilts script
 - New docs
 - New tests
-
 - Introduces FlatField class
 - Adds FlatField Notebook, tests
 - Pushes flat field algorithms into core/arflat.py
 - Main flatfield method broken into a few pieces
 - Further refactoring of armasters
 - Further refactoring related to settings and ScienceExposure
+
+- WaveImage class
+- Strip mswave from ScienceExposure
+- New tests
 
 0.7 (2017-02-07)
 ----------------